requirements.txt
# ============================================================================
# AIME Requirements - Version Optimisée pour cPanel 2.0
# Python 3.9+ - Production Ready - MySQL Compatible
# ============================================================================

# ====================
# Django Core
# ====================
Django==4.2.14
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2

# ====================
# Database - MySQL only (cPanel compatible)
# ====================
pymysql==1.1.0
mysqlclient==2.2.0

# ====================
# Static Files & Compression
# ====================
whitenoise==6.6.0

# ====================
# Security & Configuration
# ====================
python-decouple==3.8
django-cors-headers==4.3.1
django-environ==0.11.2

# ====================
# Rate Limiting & Protection
# ====================
django-ratelimit==4.1.0

# ====================
# Images & Media
# ====================
pillow==10.1.0

# ====================
# API & HTTP
# ====================
djangorestframework==3.15.2
requests==2.31.0

# ====================
# Production Server
# ====================
gunicorn==21.2.0

# ====================
# Utilities
# ====================
python-dateutil==2.8.2

# ====================
# Development Tools (commented for production)
# ====================
# django-debug-toolbar==4.2.0
# pytest==7.4.3
# pytest-django==4.6.0
